Introduce Dev only module for Kafka client extension
