+++ extensions/kafka-client/deployment/src/main/java/io/quarkus/kafka/client/deployment/KafkaProcessor.java
@@ -93,8 +93,8 @@ import io.quarkus.kafka.client.runtime.KafkaBindingConverter;
 import io.quarkus.kafka.client.runtime.KafkaRecorder;
 import io.quarkus.kafka.client.runtime.KafkaRuntimeConfigProducer;
 import io.quarkus.kafka.client.runtime.SnappyRecorder;
-import io.quarkus.kafka.client.runtime.devui.KafkaTopicClient;
-import io.quarkus.kafka.client.runtime.devui.KafkaUiUtils;
+import io.quarkus.kafka.client.runtime.dev.ui.KafkaTopicClient;
+import io.quarkus.kafka.client.runtime.dev.ui.KafkaUiUtils;
 import io.quarkus.kafka.client.runtime.graal.SnappyFeature;
 import io.quarkus.kafka.client.serialization.BufferDeserializer;
 import io.quarkus.kafka.client.serialization.BufferSerializer;
