+++ extensions/kafka-client/pom.xml
@@ -17,5 +17,6 @@
    <modules>
       <module>deployment</module>
       <module>runtime</module>
+      <module>runtime-dev</module>
    </modules>
 </project>
