+++ extensions/kafka-client/runtime-dev/src/main/java/io/quarkus/kafka/client/runtime/dev/ui/model/Order.java
@@ -1,0 +1,6 @@
+package io.quarkus.kafka.client.runtime.dev.ui.model;
+
+public enum Order {
+    OLD_FIRST,
+    NEW_FIRST
+}
