+++ extensions/kafka-client/runtime/pom.xml
@@ -86,6 +86,9 @@
                     <capabilities>
                         <provides>io.quarkus.kafka</provides>
                     </capabilities>
+                    <conditionalDevDependencies>
+                        <artifact>${project.groupId}:${project.artifactId}-dev:${project.version}</artifact>
+                    </conditionalDevDependencies>
                 </configuration>
             </plugin>
             <plugin>
