+++ extensions/kafka-client/runtime/src/main/java/io/quarkus/kafka/client/runtime/KafkaAdminClient.java
@@ -22,7 +22,6 @@ import org.apache.kafka.common.acl.AclBinding;
 import org.apache.kafka.common.acl.AclBindingFilter;
 import org.apache.kafka.common.resource.ResourcePatternFilter;
 
-import io.quarkus.kafka.client.runtime.devui.model.request.KafkaCreateTopicRequest;
 import io.smallrye.common.annotation.Identifier;
 
 @ApplicationScoped
