--- extensions/kafka-client/runtime/src/main/java/io/quarkus/kafka/client/runtime/devui/model/Order.java
@@ -1,6 +1,0 @@
-package io.quarkus.kafka.client.runtime.devui.model;
-
-public enum Order {
-    OLD_FIRST,
-    NEW_FIRST
-}
