--- extensions/kafka-client/runtime/src/main/java/io/quarkus/kafka/client/runtime/devui/model/response/KafkaMessagePage.java
@@ -1,22 +1,0 @@
-package io.quarkus.kafka.client.runtime.devui.model.response;
-
-import java.util.Collection;
-import java.util.Map;
-
-public class KafkaMessagePage {
-    private final Map<Integer, Long> nextOffsets;
-    private final Collection<KafkaMessage> messages;
-
-    public KafkaMessagePage(Map<Integer, Long> nextOffsets, Collection<KafkaMessage> messages) {
-        this.nextOffsets = nextOffsets;
-        this.messages = messages;
-    }
-
-    public Map<Integer, Long> getNextOffsets() {
-        return nextOffsets;
-    }
-
-    public Collection<KafkaMessage> getMessages() {
-        return messages;
-    }
-}
